+++ examples/run_example.py
@@ -12,10 +12,10 @@ import numpy as np
 import simtk.openmm as openmm
 import simtk.openmm.app as app
 from perses.rjmc.topology_proposal import Transformation, SamplerState
-from perses.bias_engine.base_bias_engine import BiasEngine
-from perses.alchemical_engine.base_alchemical_engine import AlchemicalEliminationEngine
+from perses.bias.bias_engine import BiasEngine
+from perses.annihilation.alchemical_engine import AlchemicalEliminationEngine
 from perses.rjmc.geometry import GeometryEngine
-from perses.ncmc_switching.base_ncmc_switching import NCMCEngine
+from perses.annihilation.ncmc_switching import NCMCEngine
 from perses.rjmc.system_engine import SystemGenerator
 
 def run():
