+++ perses/annihilation/ncmc_switching.py
